+++ src/library/hybrid_astar_lib/curve/curve_sampling_base.h
@@ -18,7 +18,7 @@ class CurveSampling {
                 const AstarRequest* request, EulerDistanceTransform* edt,
                 const ObstacleClearZone* clear_zone,
                 ParkReferenceLine* ref_line,
-                const PlannerOpenSpaceConfig* config, const float min_radius,
+                const PlannerOpenSpaceConfig* config, const double min_radius,
                 std::shared_ptr<NodeCollisionDetect> collision_detect);
 
   virtual ~CurveSampling() = default;
@@ -44,7 +44,7 @@ class CurveSampling {
   const AstarRequest *request_;
 
   const PlannerOpenSpaceConfig* config_;
-  float min_radius_;
+  double min_radius_;
   std::shared_ptr<NodeCollisionDetect> collision_detect_;
 
   // target pose for astar search.
